requirements.txt
pyyaml
scapy

black
mypy
pre-commit
